Moved glTF metadata to own directory
