+++ src/gltfMetadata/EXTMeshFeatures.ts
