+++ src/gltfMetadata/EXTStructuralMetadata.ts
