+++ src/gltfMetadata/MeshFeatures.ts
