+++ src/gltfMetadata/MeshFeaturesUtils.ts
