+++ src/gltfMetadata/StringBuilder.ts
