+++ src/gltfMetadata/StructuralMetadata.ts
